Moved the edge detection to its own file,
Added a separate file for edge detection and improved the design of the function responsible for generating the edge information. "cannyDisplay" is now named "generateEdgeImage" and simply accepts an original image and returns a new image containing the edges. Most relevant matrices are now also defined within the function instead of in the global scope.